[x64 deprecation] jax.numpy functions now default to 32-bit

+++ jax/_src/lax/control_flow.py
@@ -1857,7 +1857,7 @@ def _scan_masking_rule(padded_vals, logical_shapes, reverse, length,
   consts, init, xs = split_list(padded_vals, [num_consts, num_carry])
   max_length, = {x.shape[0] for x in xs}
   const_linear, init_linear, xs_linear = split_list(linear, [num_consts, num_carry])
-  out_vals = scan_p.bind(dynamic_length, *consts, 0, *init, *xs,
+  out_vals = scan_p.bind(dynamic_length, *consts, np.int32(0), *init, *xs,
       reverse=reverse, length=max_length, jaxpr=masked_jaxpr,
       num_consts=1 + num_consts, num_carry=1 + num_carry,
       linear=tuple([False] + const_linear + [False] + init_linear + xs_linear),
